+ CalculateController to call CalculateAPI + calculate, -form view
